html/semantics/disabled-elements/disable.button-fieldset.htm

<!DOCTYPE html>
<title>HTMLAnchorElement.text setting</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="http://www.whatwg.org/specs/web-apps/current-work/multipage/the-button-element.html#the-fieldset-element">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>


<fieldset disabled="disabled">
  <legend>test
    <button id="test_fieldset_leg_1" onclick="set_value(this)">Button</button>
  </legend>
  <button id="test_fieldset_2" onclick="set_value(this)">Button</button>
  <fieldset>
    <legend>test4
      <button id="test_fieldset_nest_leg_1" onclick="set_value(this)">Button</button>
    </legend>
    <button id="test_fieldset_nest_2" onclick="set_value(this)">Button</button>
  </fieldset>
</fieldset>


<script>
	function set_value(e){
	  var btn=document.getElementById(e.id);
	  btn.setAttribute("check","fired");
	}

	function eventFire(item, type){
	  if (item!=undefined){
	    // var evObj = document.createEvent('Events');
	    // evObj.initEvent('click', true, false);
	    // item.dispatchEvent(evObj);
	    item.click();
	  }
	}

	test(function(){
	  var btn=document.getElementById("test_fieldset_leg_1");
	  eventFire(btn,'click');
	  assert_equals(btn.getAttribute("check"),'fired');

	},'First legend event of a fieldset will fire');



	test(function(){
	  var btn=document.getElementById("test_fieldset_2");
	  eventFire(btn,'click');
	  assert_not_equals(btn.getAttribute("check"),'fired');

	},'Fieldset nested events must not fire');


	test(function(){
	  var btn=document.getElementById("test_fieldset_nest_leg_1");
	  eventFire(btn,'click');
	  assert_not_equals(btn.getAttribute("check"),'fired');

	},'Do not fire events of the first legend of a nested fieldset if that fieldset is disabled.');

	test(function(){
	  var btn=document.getElementById("test_fieldset_nest_2");
	  eventFire(btn,'click');
	  assert_not_equals(btn.getAttribute("check"),'fired');

	},'Do not fire events of a nested fieldset if that fieldset is disabled.');


	</script>
